refac(zk): move zk:value to zk/base:value
