--- tachyon/zk/BUILD.bazel
@@ -1,18 +1,0 @@
-load("//bazel:tachyon_cc.bzl", "tachyon_cc_library", "tachyon_cc_unittest")
-
-package(default_visibility = ["//visibility:public"])
-
-tachyon_cc_library(
-    name = "value",
-    hdrs = ["value.h"],
-    deps = ["//tachyon/math/base:rational_field"],
-)
-
-tachyon_cc_unittest(
-    name = "zk_unittests",
-    srcs = ["value_unittest.cc"],
-    deps = [
-        ":value",
-        "//tachyon/math/finite_fields/test:gf7",
-    ],
-)
