+++ tachyon/zk/base/value_unittest.cc
@@ -1,4 +1,4 @@
-#include "tachyon/zk/value.h"
+#include "tachyon/zk/base/value.h"
 
 #include <tuple>
